refactor: break ls into per language structure
